requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
filesystempro==0.0.0.3
idna==3.4
requests==2.31.0
urllib3==2.0.7